requirements.txt
numpy
Pillow
torch
torchvision
matplotlib
scikit-image
pyiqa
